requirements.txt
gradio
fastapi
python-dotenv
openai